+++ public/js/ec/model.js
@@ -142,6 +142,7 @@ const editActualModule = async (idModulo, modulo, inicio, fin) => {
             return $(this).text() == inicio;
           })
           .prop("selected", true);
+
         $("#MdlFin option")
           .filter(function () {
             return $(this).text() == fin;
@@ -174,7 +175,124 @@ const editActualModule = async (idModulo, modulo, inicio, fin) => {
     Swal.fire(`Error : ${error}`);
   }
 };
-
+const editActualUnitVigente = async (
+  idUnidad,
+  unidad,
+  inicio,
+  fin,
+  inicioActual,
+  finActual,
+  profesor
+) => {
+  const inicioSplit = inicio.split("/");
+  const finSplit = fin.split("/");
+  const minDate = new Date(inicioSplit[2], inicioSplit[1] - 1, inicioSplit[0]);
+  const maxDate = new Date(finSplit[2], finSplit[1] - 1, finSplit[0]);
+  try {
+    const { value: formValues } = await Swal.fire({
+      title: `Editar nombre de unidad`,
+      template: "#newUnit",
+      focusConfirm: false,
+      width: 800,
+      showLoaderOnConfirm: true,
+      showCancelButton: true,
+      allowOutsideClick: false,
+      didOpen: () => {
+        $("#UnitNombre").val(unidad);
+        $("#UnitInicio")
+          .datepicker({ dateFormat: "dd/mm/yy", minDate, maxDate })
+          .val(inicioActual);
+        $("#UnitFin")
+          .datepicker({ dateFormat: "dd/mm/yy", minDate, maxDate })
+          .val(finActual);
+        $("#UnitProfesor option")
+          .filter(function () {
+            return $(this).text() == profesor;
+          })
+          .prop("selected", true);
+      },
+      preConfirm: () => {
+        if (!$("#UnitNombre").val()) return false;
+        if (!$("#UnitProfesor").val()) return false;
+        if (!$("#UnitInicio").val()) return false;
+        if (!$("#UnitFin").val()) return false;
+        return [
+          $("#UnitNombre").val(),
+          $("#UnitProfesor").val(),
+          $("#UnitInicio").val(),
+          $("#UnitFin").val(),
+        ];
+      },
+    });
+    if (formValues) {
+      const data = {
+        unidad: formValues[0],
+        profesor: formValues[1],
+        inicio: formValues[2],
+        fin: formValues[3],
+        idUnidad,
+      };
+      const query = await $.ajax({
+        type: "PUT",
+        url: "/admin/ec/ConfigUnit",
+        data,
+      });
+      if (query) {
+        location.reload();
+      }
+    }
+  } catch (error) {
+    console.log(error);
+    Swal.fire(`Error : ${error}`);
+  }
+};
+const editActualModuleVigente = async (idModulo, modulo, inicio, fin) => {
+  try {
+    const { value: formValues } = await Swal.fire({
+      title: `Editar nombre del módulo`,
+      template: "#newModule",
+      focusConfirm: false,
+      width: 800,
+      showLoaderOnConfirm: true,
+      showCancelButton: true,
+      allowOutsideClick: false,
+      didOpen: () => {
+        $("#MdlNombre").val(modulo);
+        $("#MdlInicio").val(inicio).datepicker({ dateFormat: "dd/mm/yy" });
+        $("#MdlFin")
+          .val(fin)
+          .datepicker({ dateFormat: "dd/mm/yy", autoSize: true });
+      },
+      preConfirm: () => {
+        if (!$("#MdlNombre").val()) return false;
+        return [
+          $("#MdlNombre").val(),
+          $("#MdlInicio").val(),
+          $("#MdlFin").val(),
+        ];
+      },
+    });
+    if (formValues) {
+      const data = {
+        Nombre: formValues[0],
+        Inicio: formValues[1],
+        Fin: formValues[2],
+        idModulo,
+      };
+      const query = await $.ajax({
+        type: "PUT",
+        url: "/admin/ec/editModulo",
+        data,
+      });
+      if (query) {
+        location.reload();
+      }
+    }
+  } catch (error) {
+    console.log(error);
+    Swal.fire(`Error : ${error}`);
+  }
+};
 const deleteModulo = async (id) => {
   const alerta = await Swal.fire({
     title: "¿Eliminar el módulo?",